# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trade Request Receiver API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webhook/tradingview:
    post:
      operationId: receiveTradingViewWebhook
      tags: [webhook]
      summary: Receive webhook from TradingView
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
          text/plain:
            schema:
              type: string
      responses:
        "200":
          description: Signal received and stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeSignal"
        "400":
          description: Bad request
...
  /signals/{id}/execution:
    get:
      operationId: getSignalExecution
      tags: [executions]
      summary: Get execution record for a signal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "404":
          description: Not found
  /executions:
    get:
      operationId: listExecutions
      tags: [executions]
      summary: List recent executions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Execution"
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get current settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Save settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /settings/test-connection:
    post:
      operationId: testPublicConnection
      tags: [settings]
      summary: Test Public.com API connection
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                publicApiToken:
                  type: string
                publicAccountId:
                  type: string
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
...
components:
  schemas:
...
    WebhookPayload:
      type: object
      properties:
        ticker:
          type: ["string", "null"]
        exchange:
          type: ["string", "null"]
        interval:
          type: ["string", "null"]
        action:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        open:
          type: ["number", "null"]
        high:
          type: ["number", "null"]
        low:
          type: ["number", "null"]
        volume:
          type: ["number", "null"]
        time:
          type: ["string", "null"]
        timenow:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        basecurrency:
          type: ["string", "null"]
        quantity:
          type: ["number", "null"]
        strategy:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        position_size:
          type: ["number", "null"]
        order_price:
          type: ["number", "null"]
        order_id:
          type: ["string", "null"]
        order_comment:
          type: ["string", "null"]
    TradeSignal:
      type: object
      required: [id, ticker, action, receivedAt, raw]
      properties:
        id:
          type: integer
        ticker:
          type: string
        action:
          type: string
        price:
          type: ["number", "null"]
        open:
          type: ["number", "null"]
        high:
          type: ["number", "null"]
        low:
          type: ["number", "null"]
        volume:
          type: ["number", "null"]
        quantity:
          type: ["number", "null"]
        strategy:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        exchange:
          type: ["string", "null"]
        interval:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        basecurrency:
          type: ["string", "null"]
        alertTime:
          type: ["string", "null"]
        timenow:
          type: ["string", "null"]
        positionSize:
          type: ["number", "null"]
        orderPrice:
          type: ["number", "null"]
        orderId:
          type: ["string", "null"]
        orderComment:
          type: ["string", "null"]
        receivedAt:
          type: string
          format: date-time
        raw:
          type: object
        execution:
          $ref: "#/components/schemas/Execution"
          description: Attached execution if available
...
    Execution:
      type: object
      required: [id, signalId, status, createdAt]
      properties:
        id:
          type: integer
        signalId:
          type: integer
        status:
          type: string
          description: "pending | submitted | filled | failed | skipped"
        publicOrderId:
          type: ["string", "null"]
        orderType:
          type: ["string", "null"]
        side:
          type: ["string", "null"]
        quantity:
          type: ["string", "null"]
        limitPrice:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Settings:
      type: object
      required: [orderType, instrumentType, defaultQuantity, timeInForce, autoExecute]
      properties:
        publicAccountId:
          type: ["string", "null"]
          description: Public.com account ID
        hasApiToken:
          type: boolean
          description: Whether a Public.com API token is configured (token itself never returned)
        hasRobinhoodToken:
          type: boolean
          description: Whether a Robinhood bearer token is configured
        hasWebullKey:
          type: boolean
          description: Whether Webull App Key is configured
        webullAccountId:
          type: ["string", "null"]
          description: Webull account ID (optional)
        orderType:
          type: string
          enum: [MARKET, LIMIT]
        instrumentType:
          type: string
          enum: [EQUITY, CRYPTO]
        defaultQuantity:
          type: string
        timeInForce:
          type: string
          enum: [DAY, GTC]
        autoExecute:
          type: boolean
        buyFraction:
          type: string
          description: Fraction of defaultQuantity to use for buy orders (0.01–1.00)
        neverSellAtLoss:
          type: boolean
          description: If true, sell orders are skipped when current price is below the tracked buy price
    SettingsInput:
      type: object
      properties:
        publicApiToken:
          type: ["string", "null"]
          description: Public.com API token — pass null to keep existing
        publicAccountId:
          type: ["string", "null"]
        robinhoodBearerToken:
          type: ["string", "null"]
          description: Robinhood OAuth bearer token — pass null to keep existing
        webullAppKey:
          type: ["string", "null"]
          description: Webull App Key — pass null to keep existing
        webullAppSecret:
          type: ["string", "null"]
          description: Webull App Secret — pass null to keep existing
        webullAccountId:
          type: ["string", "null"]
        orderType:
          type: string
          enum: [MARKET, LIMIT]
        instrumentType:
          type: string
          enum: [EQUITY, CRYPTO]
        defaultQuantity:
          type: string
        timeInForce:
          type: string
          enum: [DAY, GTC]
        autoExecute:
          type: boolean
        buyFraction:
          type: string
        neverSellAtLoss:
          type: boolean
...
    ConnectionTestResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        accountId:
          type: ["string", "null"]
        accountType:
          type: ["string", "null"]
        buyingPower:
          type: ["string", "null"]
        error:
          type: ["string", "null"]